Refactor: use <Link> in Button component
This commit replaces the <a> element that's rendered in the Button
component with React Router's <Link> component. This allows us to use
client-side routing instead of doing an entire page fetch on every link
click.

This change breaks the existing tests, so we are also refactoring those a bit.
All the <Link> components rendered in the tests need to be wrapped in some kind
of router (in this case we're using BrowserRouter), otherwise the tests
break and just spit out error messages.

Instead of manually wrapping every call to render() with a
BrowserRouter, we create a testUtils.ts file that contains a helper function
that does the wrapping for us. This renderWithRouter() function adds a
BrowserRouter wrapper to the render() call, so all we need to do in the
tests is call renderWithRouter() and everything works. This function
also sets up the userEvent interface, so we'll be able to simulate user
actions such as clicking links.